address comments: finish refactor, serialize base model recursively
