+++ src/lmnr/__init__.py
@@ -1,18 +1,21 @@
-from .sdk.client.synchronous.sync_client import LaminarClient
-from .sdk.client.asynchronous.async_client import AsyncLaminarClient
-from .sdk.datasets import EvaluationDataset, LaminarDataset
-from .sdk.evaluations import evaluate
-from .sdk.laminar import Laminar
-from .sdk.types import SessionRecordingOptions, MaskInputOptions
-from .sdk.types import HumanEvaluator
-from .sdk.decorators import observe
-from .sdk.types import LaminarSpanContext
 from .opentelemetry_lib.litellm import LaminarLiteLLMCallback
 from .opentelemetry_lib.tracing.attributes import Attributes
 from .opentelemetry_lib.tracing.instruments import Instruments
 from .opentelemetry_lib.tracing.processor import LaminarSpanProcessor
 from .opentelemetry_lib.tracing.span import LaminarSpan
 from .opentelemetry_lib.tracing.tracer import get_laminar_tracer_provider, get_tracer
+from .sdk.client.asynchronous.async_client import AsyncLaminarClient
+from .sdk.client.synchronous.sync_client import LaminarClient
+from .sdk.datasets import EvaluationDataset, LaminarDataset
+from .sdk.decorators import observe
+from .sdk.evaluations import evaluate
+from .sdk.laminar import Laminar
+from .sdk.types import (
+    HumanEvaluator,
+    LaminarSpanContext,
+    MaskInputOptions,
+    SessionRecordingOptions,
+)
 
 __all__ = [
     "AsyncLaminarClient",
@@ -24,13 +27,13 @@ __all__ = [
     "LaminarClient",
     "LaminarDataset",
     "LaminarLiteLLMCallback",
+    "LaminarSpan",
     "LaminarSpanContext",
     "LaminarSpanProcessor",
-    "LaminarSpan",
+    "MaskInputOptions",
+    "SessionRecordingOptions",
+    "evaluate",
     "get_laminar_tracer_provider",
     "get_tracer",
-    "evaluate",
     "observe",
-    "SessionRecordingOptions",
-    "MaskInputOptions",
 ]
