--- src/lmnr/sdk/evaluations/consts.py
@@ -1,2 +1,0 @@
-DEFAULT_BATCH_SIZE = 5
-MAX_EXPORT_BATCH_SIZE = 64
